+++ src/lib.rs
@@ -81,4 +81,4 @@ mod links;
 mod common;
 
 pub use auth::{Token, request_token, authorize_url, authenticate_url, access_token, verify_tokens};
-pub use common::Response;
+pub use common::{Response, WebResponse};