+++ src/tweet/mod.rs
@@ -2,7 +2,6 @@
 use std::collections::HashMap;
 use auth;
 use links;
-use error;
 use common::*;
 
 mod structs;
@@ -11,7 +10,7 @@ pub use self::structs::*;
 
 ///Lookup a single tweet by numeric ID.
 pub fn show(id: i64, con_token: &auth::Token, access_token: &auth::Token)
-    -> Result<Response<Tweet>, error::Error>
+    -> WebResponse<Tweet>
 {
     let mut params = HashMap::new();
     add_param(&mut params, "id", id.to_string());