+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - New error enum, `RateLimit(i32)`, telling you that you hit your rate limit
   and when it will lapse
   - This is a **breaking change** if you were exhaustively matching on these before
+- Introduce a type alias `WebResponse` for `Result<Response<T>, Error>` which was *everywhere*
 
 ## [0.3.0] - 2016-08-19
 ### Added